requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap-modal-forms==2.2.0
django-heroku==0.3.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
